Move get dataplane functionto endpoint_context file and remove event filter predicate
